Delete unused r_cache from rutil ##api
