+++ libr/meson.build
@@ -484,7 +484,6 @@ r_util_files = [
   'include/r_util/r_big.h',
   'include/r_util/r_bitmap.h',
   'include/r_util/r_buf.h',
-  'include/r_util/r_cache.h',
   'include/r_util/r_ctypes.h',
   'include/r_util/r_debruijn.h',
   'include/r_util/r_diff.h',
