+++ libr/util/Makefile
@@ -4,7 +4,7 @@ NAME=r_util
 CFLAGS+=-DR2_PLUGIN_INCORE -I$(TOP)/shlr
 PCLIBS=@LIBZIP@ @DL_LIBS@
 OBJS=mem.o unum.o str.o hex.o file.o range.o charset.o xdg.o rxml.o
-OBJS+=prof.o cache.o sys.o buf.o sys_w32.o ubase64.o base85.o base91.o
+OBJS+=prof.o sys.o buf.o sys_w32.o ubase64.o base85.o base91.o
 OBJS+=list.o chmod.o graph.o event.o alloc.o donut.o print_code.o
 OBJS+=regex/regcomp.o regex/regerror.o regex/regexec.o uleb128.o rstr.o
 OBJS+=sandbox.o calc.o thread.o thread_sem.o thread_lock.o thread_cond.o thread_chan.o
